Cleaning up code base
