--- src/Backend/InstructorRegistry.java
@@ -1,26 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Backend;
-
-import java.util.ArrayList;
-
-/**
- *
- * @author EllieKat
- */
-public class InstructorRegistry implements Registry {
-    ArrayList users = new ArrayList();
-    
-    public InstructorRegistry(){
-        
-    }
-    
-    public boolean isInstructor(String username) {
-        
-        return true;
-    }
-    
-}
