--- src/NewBackend/Registry.java
@@ -1,15 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Backend;
-
-/**
- *
- * @author Will
- */
-public interface Registry {
-    public void writeFile();
-    public void readFile();
-}
